refactor(container): rename route views to pages
