+++ apps/container/src/pages/Admin/index.ts
@@ -1,0 +1,1 @@
+export { default as AdminPage } from './AdminPage.vue';
