+++ apps/container/src/pages/Calengrade/index.ts
@@ -1,0 +1,1 @@
+export { default as CalengradePage } from './CalengradePage.vue';
