+++ apps/container/src/pages/Confirmation/index.ts
@@ -1,0 +1,1 @@
+export { default as ConfirmationPage } from './ConfirmationPage.vue';
