+++ apps/container/src/pages/Donate/index.ts
@@ -1,0 +1,1 @@
+export { default as DonatePage } from './DonatePage.vue';
