+++ apps/container/src/pages/Help/index.ts
@@ -1,0 +1,1 @@
+export { default as HelpPage } from './HelpPage.vue';
