+++ apps/container/src/pages/History/index.ts
@@ -1,0 +1,1 @@
+export { default as HistoryPage } from './HistoryPage.vue';
