+++ apps/container/src/pages/Performance/index.ts
@@ -1,0 +1,1 @@
+export { default as PerformancePage } from './PerformancePage.vue';
