+++ apps/container/src/pages/Planning/index.ts
@@ -1,0 +1,1 @@
+export { default as PlanningPage } from './PlanningPage.vue';
