+++ apps/container/src/pages/Recovery/index.ts
@@ -1,0 +1,1 @@
+export { default as RecoveryPage } from './RecoveryPage.vue';
