+++ apps/container/src/pages/Reviews/index.ts
@@ -1,0 +1,1 @@
+export { default as ReviewsPage } from './ReviewsPage.vue';
