+++ apps/container/src/pages/Settings/index.ts
@@ -1,0 +1,1 @@
+export { default as SettingsPage } from './SettingsPage.vue';
