+++ apps/container/src/pages/Stats/index.ts
@@ -1,0 +1,1 @@
+export { default as StatsPage } from './StatsPage.vue';
