+++ apps/container/src/pages/WhatsappGroups/WhatsappGroupsPage.vue
@@ -1,0 +1,1652 @@
+<template>
+  <!-- todo: add animation -->
+  <div class="whatsapp-groups-view">
+    <UserNotifications
+      title="Tem novidade aí! 🎉"
+      text="Agora você pode filtrar os grupos por cursos para facilitar sua busca."
+    />
+    <section>
+      <div class="hero-section">
+        <h1>Encontre seus grupos do <br />Whatsapp</h1>
+        <p style="margin-bottom: 16px">
+          Acesse os grupos de WhatsApp das matérias que você está cursando e
+          fique por dentro de tudo com a sua turma.
+        </p>
+        <p style="font-size: 14px">
+          Agora você pode filtrar os grupos por cursos para facilitar sua busca.
+        </p>
+      </div>
+      <div class="search-section">
+        <div>
+          <div class="search-controls-row">
+            <div class="search-input-wrapper">
+              <Transition name="slide-up" mode="out-in">
+                <v-number-input
+                  v-if="selectedSearchType === 'ra'"
+                  v-model="searchRaQuery"
+                  placeholder="Digite seu RA (ex: 11202012345)"
+                  variant="outlined"
+                  :loading="currentLoading"
+                  :disabled="!isUserLoggedIn"
+                  prepend-inner-icon="mdi-magnify"
+                  clearable
+                  class="main-search"
+                  control-variant="hidden"
+                >
+                </v-number-input>
+
+                <div
+                  v-else-if="selectedSearchType === 'course'"
+                  class="course-search-row"
+                >
+                  <v-select
+                    v-model="searchCourseQuery"
+                    :items="coursesList"
+                    item-title="name"
+                    item-value="id"
+                    placeholder="Selecione um curso"
+                    variant="outlined"
+                    size="large"
+                    :disabled="!isUserLoggedIn"
+                    :loading="currentLoading || isCoursesLoading"
+                    prepend-inner-icon="mdi-school"
+                    class="course-select"
+                  >
+                  </v-select>
+
+                  <div class="course-filter-row">
+                    <v-text-field
+                      v-model="searchCourseFilterQuery"
+                      placeholder="Digite a disciplina ou professor"
+                      variant="outlined"
+                      size="large"
+                      :disabled="!isUserLoggedIn || !searchCourseQuery"
+                      prepend-inner-icon="mdi-magnify"
+                      clearable
+                      class="course-filter"
+                    >
+                    </v-text-field>
+
+                    <div
+                      class="season-select-wrapper season-select-wrapper--course"
+                    >
+                      <v-select
+                        v-model="selectedSeason"
+                        :items="seasonOptions"
+                        item-title="label"
+                        item-value="value"
+                        label="Quadrimestre"
+                        variant="outlined"
+                        :disabled="!isUserLoggedIn"
+                        prepend-inner-icon="mdi-calendar-clock"
+                        class="season-select"
+                      />
+                    </div>
+                  </div>
+                </div>
+
+                <v-text-field
+                  v-else
+                  v-model="searchComponentQuery"
+                  placeholder="Digite o nome da disciplina (ex: Função de várias variáveis)"
+                  variant="outlined"
+                  size="large"
+                  :disabled="!isUserLoggedIn"
+                  :loading="currentLoading"
+                  prepend-inner-icon="mdi-magnify"
+                  clearable
+                  class="main-search"
+                  @input="getWhatsappGroupsByComponent"
+                >
+                </v-text-field>
+              </Transition>
+            </div>
+            <div
+              v-if="selectedSearchType !== 'course'"
+              class="season-select-wrapper"
+            >
+              <v-select
+                v-model="selectedSeason"
+                :items="seasonOptions"
+                item-title="label"
+                item-value="value"
+                label="Quadrimestre"
+                variant="outlined"
+                :disabled="!isUserLoggedIn"
+                prepend-inner-icon="mdi-calendar-clock"
+                class="season-select"
+              />
+            </div>
+          </div>
+
+          <div class="option-chips">
+            <v-chip
+              :color="selectedSearchType === 'ra' ? 'primary' : 'default'"
+              :variant="selectedSearchType === 'ra' ? 'elevated' : 'tonal'"
+              size="large"
+              class="search-chip"
+              @click="selectSearchType('ra')"
+            >
+              <v-icon start> mdi-account </v-icon>
+              Buscar por RA
+            </v-chip>
+
+            <v-chip
+              :color="selectedSearchType === 'course' ? 'primary' : 'default'"
+              :variant="selectedSearchType === 'course' ? 'elevated' : 'tonal'"
+              size="large"
+              class="search-chip"
+              @click="selectSearchType('course')"
+            >
+              <v-icon start> mdi-school </v-icon>
+              Buscar por Curso
+            </v-chip>
+
+            <v-chip
+              :color="
+                selectedSearchType === 'component' ? 'primary' : 'default'
+              "
+              :variant="
+                selectedSearchType === 'component' ? 'elevated' : 'tonal'
+              "
+              size="large"
+              class="search-chip"
+              @click="selectSearchType('component')"
+            >
+              <v-icon start> mdi-book </v-icon>
+              Buscar por Disciplina
+            </v-chip>
+          </div>
+        </div>
+      </div>
+    </section>
+
+    <!-- Results Section -->
+    <section v-if="showOverlay" class="results-section overlay-active">
+      <div class="results-grid">
+        <WhatsappGroupCard
+          v-for="(component, index) in mockGroups"
+          :key="index"
+          :component="component"
+          class="preview-card"
+        />
+      </div>
+
+      <div class="coming-soon-overlay">
+        <div class="coming-soon-content">
+          <h2>Quase lá! 🛠️</h2>
+          <p>
+            Estamos deixando tudo pronto para 2026.2. <br />
+            Em poucos dias, você poderá acessar todos os grupos de WhatsApp das
+            suas disciplinas.
+            <br />
+            <br />
+            Enquanto isso, atualize seu histórico para garantir acesso aos
+            grupos assim que estiverem disponíveis.
+          </p>
+          <div class="not-synced__actions">
+            <button class="not-synced__button" @click="handleExtension">
+              <v-icon size="20"> mdi-link-variant </v-icon>
+              Baixar extensão
+            </button>
+            <button
+              class="not-synced__button secondary"
+              @click="handleSyncHistory"
+            >
+              <v-icon size="20"> mdi-sync </v-icon>
+              Sincronizar histórico
+            </button>
+          </div>
+        </div>
+      </div>
+    </section>
+
+    <!--  overlay if is not logged -->
+    <section v-else>
+      <div v-if="!isUserLoggedIn" class="results-success">
+        <div class="auth-required-state">
+          <div class="auth-required-visual">
+            <v-icon size="56" color="primary">mdi-lock-outline</v-icon>
+          </div>
+          <h3>Faça login para acessar os grupos de WhatsApp</h3>
+          <p>
+            Por segurança, os links dos grupos estão disponíveis apenas para
+            usuários autenticados.
+          </p>
+          <div class="not-synced__actions auth-required-actions">
+            <button class="not-synced__button" @click="createAccount">
+              <v-icon size="20"> mdi-account-plus </v-icon>
+              Criar conta
+            </button>
+            <button
+              class="not-synced__button secondary"
+              @click="handleSyncHistory"
+            >
+              <v-icon size="20"> mdi-sync </v-icon>
+              Sincronizar histórico
+            </button>
+          </div>
+        </div>
+      </div>
+
+      <div v-else-if="currentLoading" class="results-grid">
+        <v-skeleton-loader
+          v-for="(i, index) in Array.from({ length: 6 })"
+          :key="index"
+          color="secondary"
+          type="card"
+        ></v-skeleton-loader>
+      </div>
+
+      <div v-else-if="currentError" class="results-success">
+        <div class="auth-required-state">
+          <div class="auth-required-visual">
+            <v-icon size="56" color="warning">mdi-alert-circle-outline</v-icon>
+          </div>
+          <h3>Não foi possível carregar os grupos</h3>
+          <p>
+            Sua sessão pode ter expirado. Faça login novamente para continuar.
+          </p>
+          <div class="not-synced__actions">
+            <button class="not-synced__button" @click="handleLoginAgain">
+              <v-icon size="20"> mdi-login </v-icon>
+              Fazer login novamente
+            </button>
+          </div>
+        </div>
+      </div>
+
+      <div v-else-if="currentSuccess" class="results-success">
+        <div v-if="!currentGroups?.length">
+          <div class="empty-state">
+            <div class="empty-visual">
+              <v-icon size="64" color="grey-darken-1"> mdi-whatsapp </v-icon>
+            </div>
+            <h3>Nenhum grupo encontrado</h3>
+            <p class="season-warning">
+              Os grupos de {{ selectedSeasonLabel }} podem estar em atualização.
+            </p>
+
+            <div class="empty-suggestions">
+              <p class="empty-description">
+                Que tal buscar pelo nome da disciplina ou do seu curso?
+              </p>
+            </div>
+
+            <div class="empty-sync">
+              <div class="empty-sync-content">
+                <h4 class="empty-sync-title">
+                  Lembre-se de sincronizar seu histórico
+                </h4>
+                <p>
+                  Para ver os grupos das suas matérias, sincronize seu histórico
+                  de matrícula e mantenha seus dados atualizados.
+                </p>
+                <div class="not-synced__actions">
+                  <button
+                    class="not-synced__button secondary"
+                    :class="{ 'secondary--dark': isDarkMode }"
+                    @click="handleSyncHistory"
+                  >
+                    <v-icon size="20"> mdi-sync </v-icon>
+                    Sincronizar histórico
+                  </button>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        <div v-else>
+          <div class="results-grid">
+            <WhatsappGroupCard
+              v-for="(component, index) in paginatedGroups"
+              :key="index"
+              :component="component"
+              class="preview-card"
+              @open-group="openWhatsappGroup"
+            />
+          </div>
+
+          <div v-if="hasMoreResults" class="load-more-section">
+            <v-btn
+              color="primary"
+              variant="outlined"
+              size="large"
+              @click="loadMoreResults"
+            >
+              <v-icon start>mdi-chevron-down</v-icon>
+              Carregar mais ({{ currentGroups.length - paginatedGroups.length }}
+              restantes)
+            </v-btn>
+          </div>
+        </div>
+      </div>
+    </section>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { SearchCourseItem, Whatsapp } from '@next/services';
+import { useQuery } from '@tanstack/vue-query';
+import { useDebounceFn } from '@vueuse/core';
+import { computed, onMounted, ref, toValue, watch } from 'vue';
+import { useRoute, useRouter } from 'vue-router';
+import { useTheme } from 'vuetify';
+
+import UserNotifications from '@/features/whatsapp/components/UserNotifications.vue';
+import WhatsappGroupCard from '@/features/whatsapp/components/WhatsappGroupCard/WhatsappGroupCard.vue';
+import { getMockedGroups } from '@/features/whatsapp/utils/mockedWhatsappGroups';
+import { normalizeText } from '@/features/whatsapp/utils/normalizeTextSearch';
+import { eventTracker } from '@/helpers/EventTracker';
+import { WebEvent } from '@/helpers/WebEvent';
+import {
+  getCurrentAcademicSeason,
+  getSelectableAcademicSeasons,
+  isValidAcademicSeason,
+} from '@/shared/academic-calendar';
+import { useAuthStore } from '@/stores/auth';
+import { capitalizeName } from '@/utils/capitalizeName';
+import { extensionURL, studentRecordURL } from '@/utils/consts';
+import { formatSeason } from '@/utils/season';
+
+type SearchType = 'ra' | 'component' | 'course';
+const MIN_RA_LENGTH = 8;
+const DEFAULT_COURSE_NAME = 'Bacharelado em Ciências e Tecnologia';
+
+const showOverlay = ref(false);
+const defaultSeason = getCurrentAcademicSeason();
+const mockGroups = ref(getMockedGroups(defaultSeason));
+
+const router = useRouter();
+const route = useRoute();
+const authStore = useAuthStore();
+
+const handleLoginAgain = () => {
+  authStore.logOut();
+  window.location.href = '/';
+};
+
+const theme = useTheme();
+const isDarkMode = computed(() => theme.global.current.value.dark);
+
+const userRa = computed(
+  () => authStore.user?.ra || (route.query.ra ? Number(route.query.ra) : null)
+);
+const componentSelectedByQueryUrl = computed(() => {
+  const componentQuery = route.query.component;
+
+  if (Array.isArray(componentQuery)) {
+    return componentQuery[0] ?? '';
+  }
+
+  return componentQuery ?? '';
+});
+const isUserLoggedIn = computed(() => authStore.isLoggedIn);
+const isRaValid = computed(() => {
+  return (
+    searchRaQuery.value !== null &&
+    String(searchRaQuery.value).length >= MIN_RA_LENGTH
+  );
+});
+
+const searchRaQuery = ref<number | null>(null);
+const searchComponentQuery = ref('');
+const searchCourseQuery = ref<number | string | null>(null);
+const searchCourseFilterQuery = ref('');
+const routeSeason = computed(() => {
+  const seasonQuery = route.query.season;
+
+  if (Array.isArray(seasonQuery)) {
+    return seasonQuery[0] ?? '';
+  }
+
+  return String(seasonQuery ?? '');
+});
+const selectedSeason = ref(
+  isValidAcademicSeason(routeSeason.value) ? routeSeason.value : defaultSeason
+);
+const shouldFetchGroupsByRa = ref(false);
+const shouldFetchComponents = ref(false);
+const shouldFetchGroupsByCourse = ref(false);
+const selectedSearchType = ref<SearchType>('ra');
+
+const queryResultsPage = ref(1);
+
+const seasonOptions = computed(() => {
+  const currentSeason = getCurrentAcademicSeason();
+  return getSelectableAcademicSeasons(currentSeason).map((season) => ({
+    value: season,
+    label: formatSeason(season),
+  }));
+});
+
+const selectedSeasonLabel = computed(() => {
+  return formatSeason(selectedSeason.value);
+});
+
+const debouncedRaSearch = useDebounceFn((raValue: number) => {
+  if (!isUserLoggedIn.value) {
+    shouldFetchGroupsByRa.value = false;
+    return;
+  }
+
+  if (raValue && String(raValue).length >= MIN_RA_LENGTH) {
+    eventTracker.track(WebEvent.WHATSAPP_GROUP_SEARCH, {
+      search_type: 'ra',
+      search_query: raValue,
+      user_logged_in: isUserLoggedIn.value,
+    });
+
+    shouldFetchGroupsByRa.value = true;
+  } else {
+    shouldFetchGroupsByRa.value = false;
+  }
+}, 500);
+
+const applyDefaultCourseSelection = () => {
+  if (!coursesList.value.length) {
+    return;
+  }
+
+  if (
+    searchCourseQuery.value !== null &&
+    String(searchCourseQuery.value).trim()
+  ) {
+    return;
+  }
+
+  const normalizedDefault = normalizeText(DEFAULT_COURSE_NAME);
+  const defaultCourse = coursesList.value.find(
+    (course: SearchCourseItem) =>
+      normalizeText(course.name) === normalizedDefault
+  );
+
+  if (!defaultCourse) {
+    return;
+  }
+
+  searchCourseQuery.value = defaultCourse.id;
+};
+
+const selectSearchType = (type: SearchType | null) => {
+  const normalizedType = type ?? 'course';
+  if (!isUserLoggedIn.value) {
+    searchRaQuery.value = null;
+  }
+
+  searchComponentQuery.value = '';
+  searchCourseQuery.value = null;
+  searchCourseFilterQuery.value = '';
+  selectedSearchType.value = normalizedType;
+  shouldFetchGroupsByRa.value = false;
+  shouldFetchComponents.value = false;
+  shouldFetchGroupsByCourse.value = false;
+  queryResultsPage.value = 1;
+
+  if (normalizedType === 'course') {
+    applyDefaultCourseSelection();
+  }
+
+  if (normalizedType === 'ra' && isRaValid.value && searchRaQuery.value) {
+    debouncedRaSearch(searchRaQuery.value);
+  }
+};
+
+const loadMoreResults = () => {
+  queryResultsPage.value += 1;
+};
+
+const debouncedComponentSearch = useDebounceFn((query: string) => {
+  if (query.trim().length >= 2) {
+    shouldFetchComponents.value = true;
+  } else {
+    shouldFetchComponents.value = false;
+  }
+}, 300);
+
+watch(
+  componentSelectedByQueryUrl,
+  (newComponent) => {
+    if (!newComponent) {
+      return;
+    }
+
+    selectedSearchType.value = 'component';
+    searchComponentQuery.value = String(newComponent);
+    queryResultsPage.value = 1;
+    debouncedComponentSearch(String(newComponent));
+  },
+  { immediate: true }
+);
+
+watch(searchRaQuery, (newRa) => {
+  if (selectedSearchType.value === 'ra' && newRa !== null) {
+    debouncedRaSearch(newRa);
+    queryResultsPage.value = 1;
+  } else if (selectedSearchType.value === 'ra') {
+    shouldFetchGroupsByRa.value = false;
+  }
+});
+
+watch(searchComponentQuery, () => {
+  if (selectedSearchType.value === 'component') {
+    queryResultsPage.value = 1;
+  }
+});
+
+watch(searchCourseQuery, (newCourseId) => {
+  if (selectedSearchType.value === 'course' && newCourseId !== null) {
+    eventTracker.track(WebEvent.WHATSAPP_GROUP_SEARCH, {
+      search_type: 'course',
+      search_query: newCourseId,
+      user_logged_in: isUserLoggedIn.value,
+    });
+    searchCourseFilterQuery.value = '';
+    queryResultsPage.value = 1;
+    shouldFetchGroupsByCourse.value = true;
+  } else if (selectedSearchType.value === 'course') {
+    shouldFetchGroupsByCourse.value = false;
+  }
+});
+
+watch(searchCourseFilterQuery, () => {
+  if (selectedSearchType.value === 'course') {
+    queryResultsPage.value = 1;
+  }
+});
+
+watch(selectedSearchType, (newType) => {
+  if (!newType) {
+    selectSearchType('course');
+    return;
+  }
+
+  if (newType === 'course') {
+    applyDefaultCourseSelection();
+  }
+});
+
+watch(routeSeason, (newSeason) => {
+  if (!newSeason || !isValidAcademicSeason(newSeason)) {
+    return;
+  }
+
+  if (newSeason === selectedSeason.value) {
+    return;
+  }
+
+  selectedSeason.value = newSeason;
+});
+
+watch(selectedSeason, (season) => {
+  if (!isValidAcademicSeason(season)) {
+    return;
+  }
+
+  mockGroups.value = getMockedGroups(season);
+
+  const currentQuerySeason = routeSeason.value;
+  if (currentQuerySeason === season) {
+    return;
+  }
+
+  router.replace({
+    query: {
+      ...route.query,
+      season,
+    },
+  });
+});
+
+const getWhatsappGroupsByComponent = () => {
+  if (selectedSearchType.value === 'component') {
+    debouncedComponentSearch(searchComponentQuery.value);
+  }
+};
+
+// Query principal - busca todos os componentes (searchComponents retorna dados formatados do backend Next)
+const {
+  data: allComponents,
+  isLoading: isAllComponentsLoading,
+  isSuccess: isAllComponentsSuccess,
+  isError: isAllComponentsError,
+} = useQuery({
+  queryKey: ['whatsappGroups', 'allComponents', selectedSeason],
+  queryFn: () => Whatsapp.searchComponents(selectedSeason.value),
+  gcTime: 1000 * 60 * 10,
+  enabled: computed(
+    () =>
+      isUserLoggedIn.value &&
+      ['component', 'course'].includes(selectedSearchType.value)
+  ),
+  refetchOnMount: false,
+  refetchOnWindowFocus: false,
+});
+
+// Query para dados do parser (professores)
+const { data: parserComponents } = useQuery({
+  queryKey: ['disciplinaComponents', selectedSeason],
+  queryFn: () => Whatsapp.searchComponentsBySeason(selectedSeason.value),
+  gcTime: 1000 * 60 * 10,
+  enabled: computed(
+    () =>
+      isUserLoggedIn.value &&
+      ['component', 'course'].includes(selectedSearchType.value)
+  ),
+  refetchOnMount: false,
+  refetchOnWindowFocus: false,
+});
+
+// Query específica para busca por RA (necessária pois filtra no backend pelo usuário)
+const {
+  data: groupsByRa,
+  isLoading: isGroupsByRaLoading,
+  isSuccess: isGroupsByRaSuccess,
+  isError: isGroupsByRaError,
+} = useQuery({
+  queryKey: ['whatsappGroups', 'byRa', searchRaQuery, selectedSeason],
+  queryFn: () =>
+    Whatsapp.getComponentsByUser({
+      ra: searchRaQuery.value ?? 0,
+      season: selectedSeason.value,
+    }),
+  enabled: computed(() => isUserLoggedIn.value && shouldFetchGroupsByRa.value),
+});
+
+const {
+  data: coursesData,
+  isLoading: isCoursesLoading,
+  isError: isCoursesError,
+} = useQuery({
+  queryKey: ['courses'],
+  queryFn: () => Whatsapp.getCourses(),
+  enabled: computed(
+    () => isUserLoggedIn.value && selectedSearchType.value === 'course'
+  ),
+  staleTime: 1000 * 60 * 60,
+});
+
+// Componentes filtrados por nome da disciplina (busca por componente)
+const filteredByComponent = computed(() => {
+  if (!allComponentsData.value.length || !searchComponentQuery.value?.trim()) {
+    return [];
+  }
+
+  const normalizedQuery = normalizeText(searchComponentQuery.value);
+
+  return allComponentsData.value.map(enrichComponent).filter((component) => {
+    const normalizedComponentCode = normalizeText(component.uf_cod_turma || '');
+    const normalizedSubject = normalizeText(component.subject || '');
+    const normalizedCodigo = normalizeText(component.codigo || '');
+    const normalizedTeoria = normalizeText(component.teoria || '');
+    const normalizedPratica = normalizeText(component.pratica || '');
+
+    return (
+      normalizedSubject.includes(normalizedQuery) ||
+      normalizedCodigo.includes(normalizedQuery) ||
+      normalizedTeoria.includes(normalizedQuery) ||
+      normalizedPratica.includes(normalizedQuery) ||
+      normalizedComponentCode.includes(normalizedQuery)
+    );
+  });
+});
+
+const coursesList = computed(() => {
+  if (!coursesData.value) return [];
+  return coursesData.value
+    .filter((course: SearchCourseItem) => course.name && course.name.trim())
+    .sort((a: SearchCourseItem, b: SearchCourseItem) =>
+      a.name.localeCompare(b.name)
+    );
+});
+
+watch(coursesList, () => {
+  if (selectedSearchType.value === 'course') {
+    applyDefaultCourseSelection();
+  }
+});
+
+const normalizeCourseCode = (code: string | null | undefined) =>
+  String(code ?? '')
+    .trim()
+    .toUpperCase();
+
+const getBaseCourseCode = (code: string | null | undefined) =>
+  normalizeCourseCode(code).split('-')[0] ?? '';
+
+const normalizeSeasonToken = (season: string | null | undefined) =>
+  String(season ?? '')
+    .trim()
+    .replace('.', ':');
+
+const extractComponentCodeFromUfClassroom = (ufClassroomCode: string) => {
+  const match = ufClassroomCode.match(/[A-Z]{3}\d{4}-\d{2}/i);
+  return match ? normalizeCourseCode(match[0]) : '';
+};
+
+// Componentes filtrados por curso
+const filteredByCourse = computed(() => {
+  if (!allComponentsData.value.length || searchCourseQuery.value == null) {
+    return [];
+  }
+
+  const selectedCourseRaw = searchCourseQuery.value as
+    | number
+    | string
+    | { id?: number | string }
+    | null;
+  const selectedCourseId =
+    selectedCourseRaw &&
+    typeof selectedCourseRaw === 'object' &&
+    'id' in selectedCourseRaw
+      ? String(selectedCourseRaw.id ?? '').trim()
+      : String(selectedCourseRaw ?? '').trim();
+  const selectedCourse = coursesList.value.find(
+    (course: SearchCourseItem) => String(course.id).trim() === selectedCourseId
+  );
+
+  if (!selectedCourse) {
+    return [];
+  }
+
+  const courseCodes = new Set(
+    selectedCourse.ufComponentCodes.map((code) => normalizeCourseCode(code))
+  );
+  const baseCourseCodes = new Set(
+    selectedCourse.ufComponentCodes.map((code) => getBaseCourseCode(code))
+  );
+
+  const filtered = allComponentsData.value.filter((component) => {
+    const codigo = normalizeCourseCode(component.codigo);
+    const codigoBase = getBaseCourseCode(component.codigo);
+    const classroomCode = extractComponentCodeFromUfClassroom(
+      component.uf_cod_turma || ''
+    );
+    const classroomCodeBase = getBaseCourseCode(classroomCode);
+
+    const matchesCourse =
+      courseCodes.has(codigo) ||
+      baseCourseCodes.has(codigoBase) ||
+      courseCodes.has(classroomCode) ||
+      baseCourseCodes.has(classroomCodeBase);
+
+    const matchesSeason =
+      normalizeSeasonToken(component.season) ===
+      normalizeSeasonToken(selectedSeason.value);
+
+    return matchesCourse && matchesSeason;
+  });
+
+  return filtered;
+});
+
+// Componentes do curso com filtro adicional de texto e enriquecidos
+const filteredAndSearchedCourseComponents = computed(() => {
+  const courseFiltered = filteredByCourse.value;
+
+  // Enriquece os componentes com dados de professores
+  const enrichedComponents = courseFiltered.map(enrichComponent);
+
+  if (!searchCourseFilterQuery.value?.trim()) {
+    return enrichedComponents;
+  }
+
+  const normalizedQuery = normalizeText(searchCourseFilterQuery.value);
+
+  return enrichedComponents.filter((component) => {
+    const normalizedSubject = normalizeText(component.subject || '');
+    const normalizedCodigo = normalizeText(component.codigo || '');
+    const normalizedTeoria = normalizeText(component.teoria || '');
+    const normalizedPratica = normalizeText(component.pratica || '');
+
+    return (
+      normalizedSubject.includes(normalizedQuery) ||
+      normalizedCodigo.includes(normalizedQuery) ||
+      normalizedTeoria.includes(normalizedQuery) ||
+      normalizedPratica.includes(normalizedQuery)
+    );
+  });
+});
+
+// Computed para acessar dados do allComponents
+const allComponentsData = computed(() => allComponents.value || []);
+
+// Mapa de componentes com dados de professores normalizados (do parser)
+const componentsByCode = computed(() => {
+  if (!parserComponents.value) {
+    return {};
+  }
+  const grouped: Record<string, { teoria: string; pratica: string }> = {};
+
+  parserComponents.value.forEach((component) => {
+    const teoriaRaw =
+      component.teachers?.find((t: any) => t.role === 'professor')?.name || '';
+    const praticaRaw =
+      component.teachers?.find((t: any) => t.role === 'practice')?.name || '';
+
+    const teoria = capitalizeName(teoriaRaw);
+    const pratica = capitalizeName(praticaRaw);
+
+    grouped[component.ufClassroomCode] = {
+      teoria,
+      pratica,
+    };
+  });
+
+  return grouped;
+});
+
+// Função helper para enriquecer componentes com dados de professores
+const enrichComponent = (component: any) => {
+  const teachersData =
+    componentsByCode.value[component.ufClassroomCode || component.uf_cod_turma];
+  return {
+    ...component,
+    teoria: teachersData?.teoria || component.teoria || '',
+    pratica: teachersData?.pratica || component.pratica || '',
+  };
+};
+
+const groupsFromRa = computed(() => {
+  const groups = groupsByRa.value || [];
+  return groups.map(enrichComponent);
+});
+
+const searchConfig = computed(() => ({
+  ra: {
+    groups: groupsFromRa.value,
+    loading: isGroupsByRaLoading.value,
+    success: isGroupsByRaSuccess.value && shouldFetchGroupsByRa.value,
+    error: isGroupsByRaError.value,
+    query: searchRaQuery.value,
+  },
+  component: {
+    groups: filteredByComponent.value,
+    loading: isAllComponentsLoading.value,
+    success: isAllComponentsSuccess.value && shouldFetchComponents.value,
+    error: isAllComponentsError.value,
+    query: searchComponentQuery.value,
+  },
+  course: {
+    groups: filteredAndSearchedCourseComponents.value,
+    loading: isAllComponentsLoading.value || isCoursesLoading.value,
+    success: isAllComponentsSuccess.value && shouldFetchGroupsByCourse.value,
+    error: isAllComponentsError.value || isCoursesError.value,
+    query: searchCourseQuery.value,
+  },
+}));
+
+const currentGroups = computed(
+  () => searchConfig.value[selectedSearchType.value]?.groups || []
+);
+
+const paginatedGroups = computed(() => {
+  const endIndex = queryResultsPage.value * 24;
+  return currentGroups.value.slice(0, endIndex);
+});
+
+const hasMoreResults = computed(() => {
+  return paginatedGroups.value.length < currentGroups.value.length;
+});
+
+const currentLoading = computed(
+  () => searchConfig.value[selectedSearchType.value]?.loading || false
+);
+
+const currentSuccess = computed(
+  () => searchConfig.value[selectedSearchType.value]?.success || false
+);
+
+const currentError = computed(
+  () => searchConfig.value[selectedSearchType.value]?.error || false
+);
+
+watch(
+  [selectedSearchType, currentSuccess, currentLoading, currentGroups],
+  ([type, success, loading, groups]) => {
+    if (type !== 'ra' || loading || !success) {
+      return;
+    }
+
+    if (groups.length === 0) {
+      selectSearchType('course');
+    }
+  }
+);
+
+const openWhatsappGroup = (url: string) => {
+  if (!isUserLoggedIn.value || !url) {
+    return;
+  }
+
+  const component = currentGroups.value.find(
+    (group: any) => group.groupURL === url
+  );
+  const searchQuery = searchConfig.value[selectedSearchType.value]?.query;
+
+  eventTracker.track(WebEvent.WHATSAPP_GROUP_JOINED, {
+    whatsapp_url: url,
+    component: component,
+    search_type: selectedSearchType.value,
+    search_query: searchQuery,
+    user_logged_in: isUserLoggedIn.value,
+  });
+
+  window.open(url, '_blank');
+};
+
+const handleExtension = () => {
+  window.open(extensionURL, '_blank');
+};
+
+const handleSyncHistory = () => {
+  window.open(studentRecordURL, '_blank');
+};
+
+const createAccount = () => {
+  eventTracker.track(WebEvent.CREATE_ACCOUNT_CLICKED, {
+    source: 'whatsapp_groups_dialog',
+  });
+
+  router.push('/signup');
+};
+
+const openSupport = () => {
+  eventTracker.track(WebEvent.OPEN_SUPPORT, {
+    user_ra: userRa.value || null,
+    user_logged_in: isUserLoggedIn.value,
+    source: 'whatsapp_groups_dialog',
+  });
+  window.open('https://www.instagram.com/ufabc_next/?hl=pt-br', '_blank');
+};
+
+onMounted(() => {
+  eventTracker.track(WebEvent.WHATSAPP_GROUP_VIEWED, {
+    event_type: 'page_view',
+    user_logged_in: isUserLoggedIn.value,
+    user_ra: userRa.value || null,
+  });
+
+  if (isUserLoggedIn.value) {
+    if (!isValidAcademicSeason(routeSeason.value)) {
+      router.replace({
+        query: {
+          ...route.query,
+          season: selectedSeason.value,
+        },
+      });
+    }
+
+    searchRaQuery.value = toValue(userRa);
+    if (searchRaQuery.value !== null) {
+      debouncedRaSearch(searchRaQuery.value);
+    }
+  }
+});
+</script>
+
+<style scoped lang="css">
+.whatsapp-groups-view {
+  min-height: calc(100vh - 100px);
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.hero-section {
+  padding: 40px 16px 24px 16px;
+  text-align: center;
+  max-width: 600px;
+  margin: 0 auto;
+}
+
+.hero-section h1 {
+  font-size: 2.5rem;
+  font-weight: 700;
+  margin-bottom: 16px;
+  color: #2e7eed;
+}
+
+.hero-section p {
+  font-size: 1.25rem;
+  margin-bottom: 24px;
+}
+
+.not-synced__actions {
+  display: flex;
+  gap: 12px;
+  justify-content: center;
+}
+
+.hero-section {
+  text-align: center;
+}
+
+.search-section {
+  padding: 0 16px;
+  max-width: 800px;
+  margin: 0 auto;
+  margin-bottom: 16px;
+}
+
+.search-controls-row {
+  display: flex;
+  gap: 12px;
+  align-items: stretch;
+  margin-bottom: 12px;
+}
+
+.season-select-wrapper {
+  flex: 1;
+  min-width: 180px;
+}
+
+.search-input-wrapper {
+  flex: 3;
+}
+
+.season-select :deep(.v-field) {
+  border-radius: 16px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.main-search :deep(.v-field) {
+  border-radius: 16px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.course-search-row {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  width: 100%;
+}
+
+.course-filter-row {
+  display: flex;
+  gap: 16px;
+  width: 100%;
+}
+
+.course-select {
+  flex: 1.5;
+  min-width: 0;
+}
+
+.course-select :deep(.v-field) {
+  border-radius: 16px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.course-filter {
+  flex: 3;
+  min-width: 0;
+}
+
+.course-filter :deep(.v-field) {
+  border-radius: 16px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.season-select-wrapper--course {
+  flex: 1;
+}
+
+@media (max-width: 768px) {
+  .search-controls-row {
+    flex-direction: column;
+  }
+
+  .season-select-wrapper,
+  .search-input-wrapper {
+    width: 100%;
+    flex: 1;
+  }
+
+  .season-select-wrapper {
+    order: -1;
+  }
+
+  .course-search-row {
+    flex-direction: column;
+    gap: 12px;
+  }
+
+  .course-filter-row {
+    flex-direction: column;
+    gap: 12px;
+  }
+
+  .course-select {
+    flex: 1;
+  }
+
+  .course-filter,
+  .season-select-wrapper--course {
+    flex: 1;
+    width: 100%;
+  }
+}
+
+.option-chips {
+  display: flex;
+  gap: 12px;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+.search-chip {
+  transition: all 0.2s ease;
+}
+
+.search-chip:hover:not(:disabled) {
+  transform: translateY(-2px);
+}
+
+.results-section {
+  position: relative;
+  padding: 32px 16px;
+  max-width: 1200px;
+  min-height: 400px;
+}
+
+.results-success {
+  padding: 32px 16px;
+  max-width: 1200px;
+  margin: 0 auto;
+  min-height: 400px;
+}
+
+.auth-required-state {
+  margin: 0 auto;
+  max-width: 640px;
+  text-align: center;
+  padding: 16px;
+}
+
+.auth-required-visual {
+  margin-bottom: 16px;
+}
+
+.auth-required-state h3 {
+  margin-bottom: 12px;
+  color: rgb(var(--v-theme-on-surface));
+}
+
+.auth-required-state p {
+  margin-bottom: 20px;
+}
+
+.season-warning {
+  margin-top: -4px;
+  margin-bottom: 16px;
+  color: #b45309;
+  font-weight: 500;
+}
+
+.coming-soon-overlay {
+  position: absolute;
+  height: 100%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: linear-gradient(
+    to bottom,
+    rgba(255, 255, 255, 0) 0%,
+    rgba(255, 255, 255, 0.8) 30%,
+    rgba(255, 255, 255, 0.95) 60%,
+    rgba(255, 255, 255, 1) 100%
+  );
+  backdrop-filter: blur(2px);
+  border-radius: 0 0 20px 20px;
+  display: flex;
+  align-items: end;
+  justify-content: center;
+  z-index: 10;
+}
+
+.coming-soon-content {
+  text-align: center;
+  color: #2e7eed;
+  max-width: 500px;
+  padding-bottom: 20px;
+}
+
+.coming-soon-content h2 {
+  font-size: 2.5rem;
+  font-weight: 700;
+  margin-bottom: 16px;
+  color: #2e7eed;
+  text-shadow: none;
+}
+
+.coming-soon-content p {
+  font-size: 1.2rem;
+  margin-bottom: 32px;
+  line-height: 1.6;
+  color: #2d2d2d;
+  opacity: 1;
+}
+
+.coming-soon-features {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  margin-bottom: 32px;
+}
+
+.feature-item {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  font-size: 1.1rem;
+  background: rgba(46, 126, 237, 0.15);
+  padding: 12px 20px;
+  border-radius: 12px;
+  color: #2e7eed;
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(46, 126, 237, 0.2);
+}
+
+.empty-state {
+  text-align: center;
+  padding: 24px 16px;
+}
+
+.empty-suggestions {
+  margin-top: 24px;
+}
+
+.empty-description {
+  margin-bottom: 32px;
+  font-size: 1.1rem;
+  max-width: 600px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.empty-sync {
+  margin: 0 auto;
+  max-width: 600px;
+  padding: 20px 18px;
+  border-radius: 16px;
+  background-color: var(--v-theme-primary);
+  border: 1px solid rgba(var(--v-theme-primary), 0.1);
+}
+
+.empty-sync-content {
+  text-align: center;
+  color: rgb(var(--v-theme-primary));
+}
+
+.empty-sync-content h4 {
+  font-size: 1.25rem;
+  font-weight: 600;
+  margin-bottom: 10px;
+  color: rgb(var(--v-theme-primary));
+}
+
+.empty-sync-title {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.empty-sync-content p {
+  font-size: 1rem;
+  line-height: 1.5;
+  color: rgb(var(--v-theme-on-surface));
+  margin-bottom: 16px;
+}
+
+.empty-visual {
+  margin-bottom: 24px;
+}
+
+.empty-state h3 {
+  margin-bottom: 8px;
+  color: rgb(var(--v-theme-on-surface));
+}
+
+.empty-state p {
+  margin-bottom: 24px;
+}
+
+.results-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
+  gap: 20px;
+  position: relative;
+  z-index: 1;
+}
+
+.preview-card {
+  animation: fadeInUp 0.8s ease-out forwards;
+}
+
+.results-section.overlay-active .preview-card {
+  filter: blur(1px);
+  pointer-events: none;
+}
+
+.not-synced__button {
+  background-color: rgb(var(--v-theme-primary));
+  color: white;
+  border-radius: 8px;
+  padding: 12px 24px;
+  font-size: 16px;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+.not-synced__button.secondary {
+  background-color: white;
+  color: rgb(var(--v-theme-primary));
+  border: 1px solid rgb(var(--v-theme-primary));
+}
+
+.not-synced__button.secondary--dark {
+  background-color: rgb(var(--v-theme-surface));
+  color: rgb(var(--v-theme-primary));
+  border: 1px solid rgb(var(--v-theme-primary));
+}
+
+.not-synced__button:hover {
+  transform: translateY(-4px);
+  transition: all 0.2s ease;
+}
+
+.auth-required-actions .not-synced__button:hover {
+  transform: translateY(-1px);
+}
+
+/* vue transition */
+.slide-up-enter-active,
+.slide-up-leave-active {
+  transition: all 0.25s ease-out;
+}
+
+.slide-up-enter-from {
+  opacity: 0;
+  transform: translateY(30px);
+}
+
+.slide-up-leave-to {
+  opacity: 0;
+  transform: translateY(-30px);
+}
+
+.link-style {
+  text-decoration: none;
+  color: #37bba3;
+  cursor: pointer;
+}
+
+@keyframes fadeInUp {
+  from {
+    opacity: 0;
+    transform: translateY(30px);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@media (max-width: 768px) {
+  .coming-soon-content h2 {
+    font-size: 2rem;
+  }
+
+  .coming-soon-content p {
+    font-size: 1.1rem;
+  }
+
+  .feature-item {
+    font-size: 1rem;
+    padding: 10px 16px;
+  }
+}
+
+/* Dialog Styles - Inspired by DonatePage */
+.restriction-dialog {
+  border-radius: 16px !important;
+  padding: 20px;
+  max-height: 90vh;
+  overflow-y: auto;
+}
+
+.dialog-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 20px;
+  position: sticky;
+  top: 0;
+  z-index: 1;
+  padding-bottom: 8px;
+}
+
+.dialog-title {
+  font-weight: 500;
+  font-size: 1.5rem;
+  color: rgb(var(--v-theme-primary));
+  padding: 0;
+  font-family: 'Lato', 'Roboto';
+  line-height: 1.2;
+  flex: 1;
+  margin-right: 16px;
+}
+
+.dialog-close-btn {
+  padding: 0;
+  min-height: auto;
+  flex-shrink: 0;
+}
+
+.dialog-content {
+  padding: 0 0 24px 0;
+  line-height: 1.6;
+}
+
+.main-message {
+  margin-bottom: 24px;
+}
+
+.main-message p {
+  font-size: 1.1rem;
+  color: rgb(var(--v-theme-on-surface));
+  margin: 0;
+}
+
+.info-section,
+.recommendation-section,
+.recommendations-section {
+  background: rgba(var(--v-theme-primary), 0.05);
+  border-radius: 12px;
+  padding: 20px;
+  margin-bottom: 20px;
+  border-left: 4px solid rgb(var(--v-theme-primary));
+}
+
+.info-section h4,
+.recommendation-section h4,
+.recommendations-section h4 {
+  color: rgb(var(--v-theme-primary));
+  font-weight: 600;
+  margin-bottom: 12px;
+  font-size: 1.1rem;
+}
+
+.info-section p,
+.recommendation-section p,
+span {
+  margin: 0;
+  color: rgb(var(--v-theme-on-surface));
+  font-size: 1rem;
+}
+
+.safety-tips {
+  margin: 0;
+  padding-left: 20px;
+  color: rgb(var(--v-theme-on-surface));
+  font-size: 1rem;
+}
+
+.safety-tips li {
+  margin-bottom: 8px;
+}
+
+.community-message {
+  background: linear-gradient(
+    135deg,
+    rgba(var(--v-theme-primary), 0.1) 0%,
+    rgba(var(--v-theme-primary), 0.05) 100%
+  );
+  border-radius: 12px;
+  padding: 20px;
+  text-align: center;
+}
+
+.community-message p {
+  margin: 0;
+  font-size: 1rem;
+  color: rgb(var(--v-theme-primary));
+  line-height: 1.5;
+}
+
+.dialog-actions {
+  padding: 0;
+  justify-content: center;
+}
+
+.confirm-btn {
+  font-family: 'Roboto';
+  color: white !important;
+  width: 100%;
+  height: 48px;
+  font-size: 1.1rem;
+  font-weight: 500;
+}
+
+@media (max-width: 600px) {
+  .restriction-dialog {
+    margin: 0;
+    padding: 16px;
+    border-radius: 0 !important;
+    max-height: 100vh;
+    height: 100vh;
+  }
+
+  .dialog-header {
+    margin-bottom: 16px;
+    padding: 0 0 12px 0;
+    border-bottom: 1px solid rgba(var(--v-theme-primary), 0.1);
+  }
+
+  .dialog-title {
+    font-size: 1.2rem;
+    line-height: 1.3;
+    margin-right: 12px;
+  }
+
+  .dialog-content {
+    padding: 0 0 20px 0;
+  }
+
+  .main-message {
+    margin-bottom: 20px;
+  }
+
+  .main-message p {
+    font-size: 0.95rem;
+    line-height: 1.5;
+  }
+
+  .info-section,
+  .recommendation-section,
+  .recommendations-section {
+    padding: 14px;
+    margin-bottom: 16px;
+    border-radius: 8px;
+  }
+
+  .info-section h4,
+  .recommendation-section h4,
+  .recommendations-section h4 {
+    font-size: 1rem;
+    margin-bottom: 10px;
+  }
+
+  .info-section p,
+  .recommendation-section p {
+    font-size: 0.9rem;
+    line-height: 1.4;
+  }
+
+  .safety-tips {
+    font-size: 0.9rem;
+    line-height: 1.4;
+    padding-left: 16px;
+  }
+
+  .safety-tips li {
+    margin-bottom: 6px;
+  }
+
+  .community-message {
+    padding: 14px;
+    border-radius: 8px;
+  }
+
+  .community-message p {
+    font-size: 0.9rem;
+    line-height: 1.4;
+  }
+
+  .dialog-actions {
+    padding: 16px 0 0 0;
+    position: sticky;
+    bottom: 0;
+    background: rgb(var(--v-theme-surface));
+    border-top: 1px solid rgba(var(--v-theme-primary), 0.1);
+  }
+
+  .confirm-btn {
+    height: 44px;
+    font-size: 1rem;
+  }
+}
+
+/* Adicionar scroll suave para o dialog */
+.restriction-dialog {
+  scroll-behavior: smooth;
+}
+
+.restriction-dialog::-webkit-scrollbar {
+  width: 4px;
+}
+
+.restriction-dialog::-webkit-scrollbar-track {
+  background: rgba(var(--v-theme-primary), 0.1);
+  border-radius: 2px;
+}
+
+.restriction-dialog::-webkit-scrollbar-thumb {
+  background: rgba(var(--v-theme-primary), 0.3);
+  border-radius: 2px;
+}
+
+.restriction-dialog::-webkit-scrollbar-thumb:hover {
+  background: rgba(var(--v-theme-primary), 0.5);
+}
+
+.load-more-section {
+  display: flex;
+  justify-content: center;
+  padding: 32px 0;
+  margin-top: 16px;
+}
+</style>
