+++ apps/container/src/pages/WhatsappGroups/index.ts
@@ -1,0 +1,1 @@
+export { default as WhatsappGroupsPage } from './WhatsappGroupsPage.vue';
