+++ apps/container/src/router/index.ts
@@ -1,6 +1,5 @@
 import { Auth } from '@next/services';
 import type { RouteLocationNormalized, RouteRecordRaw } from 'vue-router';
-import { logger } from '@/utils/logger';
 import { createRouter, createWebHistory } from 'vue-router';
 
 import {
@@ -12,29 +11,30 @@ import {
 } from '@/router/auth/authConfig';
 import { useAuthStore } from '@/stores/auth';
 import { isUserTokenExpired, isValidJwtFormat } from '@/utils/jwt';
+import { logger } from '@/utils/logger';
 
-const ReviewsView = () => import('@/views/Reviews/ReviewsView.vue');
-const PerformanceView = () => import('@/views/Performance/PerformanceView.vue');
-const PlanningView = () => import('@/views/Planning/PlanningView.vue');
-const HistoryView = () => import('@/views/History/HistoryView.vue');
-const StatsView = () => import('@/views/Stats/StatsView.vue');
-const SettingsView = () => import('@/views/Settings/SettingsView.vue');
-const DonateView = () => import('@/views/Donate/DonateView.vue');
-const SignUpView = () => import('@/views/SignUp/SignUpView.vue');
-const ConfirmationView = () =>
-  import('@/views/Confirmation/ConfirmationView.vue');
-const RecoveryView = () => import('@/views/Recovery/RecoveryView.vue');
-const LoginView = () => import('@/views/Login/LoginView.vue');
-const CalengradeView = () => import('@/views/Calengrade/CalengradeView.vue');
-const WhatsappGroupsView = () =>
-  import('@/views/WhatsappGroups/WhatsappGroupsView.vue');
-const HelpView = () => import('@/views/Help/HelpView.vue');
+const ReviewsPage = () => import('@/pages/Reviews/ReviewsPage.vue');
+const PerformancePage = () => import('@/pages/Performance/PerformancePage.vue');
+const PlanningPage = () => import('@/pages/Planning/PlanningPage.vue');
+const HistoryPage = () => import('@/pages/History/HistoryPage.vue');
+const StatsPage = () => import('@/pages/Stats/StatsPage.vue');
+const SettingsPage = () => import('@/pages/Settings/SettingsPage.vue');
+const DonatePage = () => import('@/pages/Donate/DonatePage.vue');
+const SignUpPage = () => import('@/pages/SignUp/SignUpPage.vue');
+const ConfirmationPage = () =>
+  import('@/pages/Confirmation/ConfirmationPage.vue');
+const RecoveryPage = () => import('@/pages/Recovery/RecoveryPage.vue');
+const LoginPage = () => import('@/pages/Login/LoginPage.vue');
+const CalengradePage = () => import('@/pages/Calengrade/CalengradePage.vue');
+const WhatsappGroupsPage = () =>
+  import('@/pages/WhatsappGroups/WhatsappGroupsPage.vue');
+const HelpPage = () => import('@/pages/Help/HelpPage.vue');
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/reviews',
     name: 'reviews',
-    component: ReviewsView,
+    component: ReviewsPage,
     meta: {
       title: 'Reviews',
       requiresAuth: true,
@@ -44,7 +44,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/performance',
     name: 'performance',
-    component: PerformanceView,
+    component: PerformancePage,
     meta: {
       title: 'Performance',
       requiresAuth: true,
@@ -54,7 +54,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/planning',
     name: 'planning',
-    component: PlanningView,
+    component: PlanningPage,
     meta: {
       title: 'Planejamento',
       requiresAuth: true,
@@ -64,7 +64,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/history',
     name: 'history',
-    component: HistoryView,
+    component: HistoryPage,
     meta: {
       title: 'Meu Histórico',
       requiresAuth: true,
@@ -74,7 +74,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/stats',
     name: 'stats',
-    component: StatsView,
+    component: StatsPage,
     meta: {
       title: 'Dados da Matrícula',
       requiresAuth: true,
@@ -84,7 +84,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/settings',
     name: 'settings',
-    component: SettingsView,
+    component: SettingsPage,
     meta: {
       title: 'Configurações',
       requiresAuth: true,
@@ -94,7 +94,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/donate',
     name: 'donate',
-    component: DonateView,
+    component: DonatePage,
     meta: {
       title: 'Ajude o Next',
       layout: 'include-sidebar',
@@ -103,7 +103,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     name: 'signup',
     path: '/signup',
-    component: SignUpView,
+    component: SignUpPage,
     meta: {
       title: 'Cadastro',
       unconfirmedOnly: true,
@@ -113,7 +113,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     name: 'confirm',
     path: '/confirm',
-    component: ConfirmationView,
+    component: ConfirmationPage,
     meta: {
       title: 'Confirmação da conta',
       unconfirmedOnly: true,
@@ -122,7 +122,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/recovery',
     name: 'recovery',
-    component: RecoveryView,
+    component: RecoveryPage,
     meta: {
       title: 'Recuperar conta',
       guestOnly: true,
@@ -131,7 +131,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/login',
     name: 'login',
-    component: LoginView,
+    component: LoginPage,
     meta: {
       title: 'Entrar no Next',
       guestOnly: true,
@@ -140,7 +140,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/grupos-whatsapp',
     name: 'whatsapp',
-    component: WhatsappGroupsView,
+    component: WhatsappGroupsPage,
     meta: {
       title: 'Grupos do Whatsapp',
       layout: 'include-sidebar',
@@ -149,7 +149,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/calengrade',
     name: 'calengrade',
-    component: CalengradeView,
+    component: CalengradePage,
     meta: {
       title: 'Calengrade',
       layout: 'include-sidebar',
@@ -159,7 +159,7 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/help',
     name: 'help',
-    component: HelpView,
+    component: HelpPage,
     meta: {
       title: 'Ajuda',
       layout: 'include-sidebar',
@@ -290,14 +290,14 @@ function resolveRouteAccess(to: RouteLocationNormalized) {
   const authenticatedRedirectPath = AUTHENTICATED_REDIRECT_PATH;
 
   const requiresAuth = to.matched.some(
-    (record) => record.meta.requiresAuth === true,
+    (record) => record.meta.requiresAuth === true
   );
   const requiresConfirmed = to.matched.some(
-    (record) => record.meta.requiresConfirmed === true,
+    (record) => record.meta.requiresConfirmed === true
   );
   const guestOnly = to.matched.some((record) => record.meta.guestOnly === true);
   const unconfirmedOnly = to.matched.some(
-    (record) => record.meta.unconfirmedOnly === true,
+    (record) => record.meta.unconfirmedOnly === true
   );
 
   if (requiresConfirmed) {
