--- apps/container/src/views/Admin/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminView } from './AdminView.vue';
