--- apps/container/src/views/Calengrade/index.ts
@@ -1,1 +1,0 @@
-export { default as CalengradeVue } from './CalengradeView.vue';
