--- apps/container/src/views/Confirmation/index.ts
@@ -1,1 +1,0 @@
-export { default as ConfirmationView } from './ConfirmationView.vue';
