--- apps/container/src/views/Donate/index.ts
@@ -1,1 +1,0 @@
-export { default as DonateView } from './DonateView.vue';
