--- apps/container/src/views/Help/index.ts
@@ -1,1 +1,0 @@
-export { default as HelpView } from './HelpView.vue';
