--- apps/container/src/views/History/index.ts
@@ -1,1 +1,0 @@
-export { default as HistoryView } from './HistoryView.vue';
