--- apps/container/src/views/Login/index.ts
@@ -1,1 +1,0 @@
-export { default as LoginView } from './LoginView.vue';
