--- apps/container/src/views/Performance/index.ts
@@ -1,1 +1,0 @@
-export { default as PerformanceView } from './PerformanceView.vue';
