--- apps/container/src/views/Planning/index.ts
@@ -1,1 +1,0 @@
-export { default as PlanningView } from './PlanningView.vue';
