--- apps/container/src/views/Recovery/index.ts
@@ -1,1 +1,0 @@
-export { default as RecoveryView } from './RecoveryView.vue';
