--- apps/container/src/views/Reviews/index.ts
@@ -1,1 +1,0 @@
-export { default as ReviewsView } from './ReviewsView.vue';
