--- apps/container/src/views/Settings/index.ts
@@ -1,1 +1,0 @@
-export { default as SettingsView } from './SettingsView.vue';
