--- apps/container/src/views/SignUp/index.ts
@@ -1,1 +1,0 @@
-export { default as SignUpView } from './SignUpView.vue';
