--- apps/container/src/views/Stats/index.ts
@@ -1,1 +1,0 @@
-export { default as StatsView } from './StatsView.vue';
